move the cart-pole speed test to a unit test
